==> code_updated/02_Create_Datalist-FLAIRONLY.ipynb <==
# Create JSON data list 

In this notebook, I want to create the JSON datalist as exemplified by the rest example_config_datalist: 

import os
import glob
import json
from pathlib import Path
import pandas as pd
import subprocess as sub 
from subprocess import Popen, PIPE, STDOUT

# Example file: 

with open('/working/lupolab/julia/tcia_analysis/nvidia/config/config_fla-t2all_datalist.json') as json_file:
    example_list = json.load(json_file)

example_list

## Need to mimic the structure of this 'validation' list 
# example_list['validation']

## It is a list of dictionaries that have 'label' keys and 'image' keys; label key will be left blank 
## Image key will be list of stirngs describing nifti input 
## Must be in the following order: 1) t1c 2) t1 3) t2 4) t2flair 

## Use results from labeling (that we did during registration)

PATH = Path('/working/lupolab/julia/tcia_analysis/datasets/')

full_label_df = pd.read_csv(Path(PATH, 'one_exam_per_patient_with_labels.csv'), index_col = 0)
full_label_df = full_label_df.reset_index(drop = True)

full_label_df.shape

root_dir = '/working/lupolab/julia/tcia_analysis'

patient_id = "TCGA-06-1801"

full_label_df.image[0]

full_label_df.head()

full_label_df.contrast_label.value_counts()

flair_df = full_label_df.loc[full_label_df.contrast_label == "T2_FLAIR"]

flair_df.shape

flair_df_nodup = flair_df.drop_duplicates(subset = ['patient_id'])

flair_df_nodup.shape

flair_df_nodup.head()

valid_list = []
for idx, row in flair_df_nodup.iterrows(): 
    flair_image_loc = row['cohort']+"-nifti/"+row['patient_id']+"/"+row['patient_exam']+"/"+row['image']+'a.nii.gz'
    full_path = Path(PATH, Path(flair_image_loc).parent)
#     print(full_path)
    if os.path.isdir(full_path):
        if flair_image_loc.split('/')[-1] in os.listdir(full_path) :
            flair_dict = {'image': flair_image_loc}
            valid_list.append(flair_dict)
    else: print(full_path)
    

len(valid_list)

validation_datalist = {'training': [], 'validation': valid_list}

validation_datalist

with open('/working/lupolab/julia/tcia_analysis/nvidia/config/config_datalist_flair.json', 'w') as fp:
    json.dump(validation_datalist, fp)
