==> code_updated/03_Segmentation_Commands.ipynb <==
import os
import glob
import json
from pathlib import Path
import pandas as pd
import subprocess as sub 
from subprocess import Popen, PIPE, STDOUT

# 1. Change the infer.sh script

PATH = Path('/working/lupolab/julia/tcia_analysis/')

os.listdir(Path(PATH, 'nvidia/commands'))

Copy `infer.sh to` `infer_t1c.sh`

`nano infer_t1c.sh`


1. Change `ENVIRONMENT_FILE=config/environment.json` to `ENVIRONMENT_FILE=config/environment_t1c.json`
2. Change `DATASET_JSON=/working/lupolab/julia/tcia_analysis/nvidia/config/config_datalist.json` to `DATASET_JSON=/working/lupolab/julia/tcia_analysis/nvidia/config/config_datalist_t1c.json`

# 2. Change config files: 

## 2.1 environment.json

Change `"DATASET_JSON": "/working/lupolab/julia/tcia_analysis/nvidia/config/config_datalist_flair.json"` to `"DATASET_JSON": "/working/lupolab/julia/tcia_analysis/nvidia/config/config_datalist_t1c.json"`

Change `"MMAR_EVAL_OUTPUT_PATH": "eval_t1c"`

# 3. Run inference: 

`singularity exec --nv -B /working/lupolab/julia/tcia_analysis:/working/lupolab/julia/tcia_analysis -B /scratch/sf673542:/scratch --pwd $PWD /data/i2/nvidia/singularity/clara-train-sdk-v2-api.img ./infer_t1c.sh`

Note, stalled on index 9 (batch 10) so I got rid of i
