==> code_updated/04_FLAIR-map_segs_to_PID_visuallyVerify-redo.ipynb <==
import os 
import glob
from pathlib import Path
import pandas as pd 
import numpy as np 
import nibabel as nib 
import json
from nilearn import plotting, image
import matplotlib.pyplot as plt
import matplotlib.cm as cm

pd.set_option('display.max_colwidth', 1000)

# 1 Map the segmentations (named for fla image) to their corresponding patient: 

Begin by reading in the json file and data frames that describe what we just segmented: 

PATH = Path('/working/lupolab/julia/tcia_analysis/datasets/')

df = pd.read_csv(Path(PATH, 'one_exam_per_patient_with_fixed_vol.csv'))

with open(Path(PATH.parent,'nvidia', 'config', 'config_datalist_flair.json'), 'r') as f: 
    segment_json = json.load(f)

segment_json

Now list all of the segmentations that are in our 'eval_fla' folder: 

segmentations = [Path(x).stem for x in glob.glob(str(Path(PATH.parent, 'nvidia', 'eval_fla', '*a'))) if x not in glob.glob(str(Path(PATH.parent, 'nvidia', 'eval_fla', '*flaa')))]

len(segmentations)

segmentations[:10]

df.head()

df.shape

df_nodup = df.drop_duplicates(subset=['patient_id'])
df_nodup = df_nodup.drop(['image', 'contrast', 'contrast_label', 'plane', 'exam_eligible', 'fixed_vol'], axis = 1)

df_nodup.shape

df_nodup = df_nodup.reset_index(drop = True)

df_nodup.head()

seg_loc_root = Path(PATH.parent, 'nvidia', 'eval_fla')

seg_list = [x['image'] for x in segment_json['validation']]

seg_list

attempted_segmentations = pd.DataFrame({'patient_id': [x.split('/')[1] for x in seg_list], 
                                        'cohort': [x.split('/')[0] for x in seg_list], 
                                        'patient_exam': [x.split('/')[2] for x in seg_list], 
                                        
             'seg_loc': [x.split('/')[3].split('.')[-3] for x in seg_list], 
             'seg_file': [ x.split('/')[3].split('.')[-3]+'_TC.nii.gz' for x in seg_list]})

actual_segmentations = pd.DataFrame(segmentations)

actual_segmentations.shape, attempted_segmentations.shape

actual_segmentations.head()

attempted_segmentations.head()

actual_seg_map = attempted_segmentations.loc[attempted_segmentations.seg_loc.isin(list(actual_segmentations[0]))]

actual_seg_map.shape

actual_seg_map['fla'] = [x[:-10]+'.nii.gz' for x in actual_seg_map.seg_file]

actual_seg_map.head()

tcia_df = actual_seg_map

# 2 Shift the segmentations to match the original header 

base_dest_path = '/working/lupolab/julia/tcia_analysis/nvidia/eval_fla/'

PATH_TCIA = Path("/working/lupolab/julia/tcia_analysis")

Find the original paths for the images: 

fla_file_paths = []
for idx, row in tcia_df.iterrows():
    if type(row['fla']) != float: 
        orig_fla_file_path = str(Path(PATH_TCIA, 'datasets',row['cohort'], row['patient_id'], row['patient_exam'],  row['fla']))
        fla_file_paths.append(orig_fla_file_path)

os.chdir(base_dest_path)

# !rm -r *_fla

Read in the original image and the segmentation, shift the segmentation & save it in the same folder as the segmentation: 

from nilearn.image import new_img_like

for idx, row in actual_seg_map.iterrows():
    original_file     = Path(PATH_TCIA, 'datasets', row['cohort'], row['patient_id'], row['patient_exam'], row['fla'])
    segmentation_file = Path(PATH_TCIA, 'nvidia', 'eval_fla', row['seg_loc'], row['seg_file']) 
    orig_fla_nii      = nib.load(original_file)
    seg_fla_nii       = nib.load(segmentation_file)
    shifted_nifti     = new_img_like(orig_fla_nii, 
                                     seg_fla_nii.get_fdata(), 
                                     affine=orig_fla_nii.affine, 
                                     copy_header=True)
    # Ensuring that the shapes match: 
    if shifted_nifti.shape == orig_fla_nii.shape:
        print(f"Shapes match for {row['patient_id']}")
    else: print(f"Shapes don't match for {row['patient_id']}") 
    
    # Save the shifted nifti in the folder where the original segmentation was: 
    shifted_seg_name = Path(segmentation_file.parent, row['seg_file'][:-7]+'_shifted'+row['seg_file'][-7:])
    print(f'Saving {shifted_seg_name}')
    nib.save(shifted_nifti, Path(segmentation_file.parent, shifted_seg_name))

## 2.2 Check to make sure that all the segmentations have a corresponding shifted segmentation: 

segmentations = glob.glob(base_dest_path+'/*/*a_TC.nii.gz')
len(segmentations)

segmentations = glob.glob(base_dest_path+'/*/*a_TC_shifted.nii.gz')
len(segmentations)

## 2.3 Change the data frame to point to the shifted segmentations instead of the original ones: 

actual_seg_map['seg_file_shifted'] = [x[:-7]+'_shifted'+x[-7:] for x in actual_seg_map.seg_file]

actual_seg_map = actual_seg_map.drop(['seg_file'], axis = 1)


actual_seg_map.head()

actual_seg_map.columns = ['patient_id', 'cohort', 'patient_exam', 'seg_loc', 'fla', 'seg_file']

seg_df = actual_seg_map

seg_df['seg_loc'] = [str(Path(seg_loc_root, x)) for x in list(seg_df['seg_loc'])]

seg_df.head()

seg_df.to_csv(Path(PATH, 'segmentation_fla_df.csv'), index = False)

os.listdir(PATH)

# 3 Write out the pngs to ensure that they look ok: 

## Let's investigate just a few of them to ensure that they are correct: 
import nibabel as nib
import matplotlib.pyplot as plt

%matplotlib inline

plotting_dict = {'display_mode':'z', 
                 'cut_coords':[-50,-30,-10, 10, 30, 50], 
                 'cmap':cm.get_cmap('viridis')}

for idx, row in seg_df.iterrows(): 
    
    original_file     = Path(PATH_TCIA, 'datasets', row['cohort'], row['patient_id'], row['patient_exam'], row['fla'])
    segmentation_file = Path(PATH_TCIA, 'nvidia', 'eval_fla', row['seg_loc'], row['seg_file']) 
    orig_fla_nii      = nib.load(original_file)
    seg_fla_nii       = nib.load(segmentation_file)
    
    png_path = Path(base_dest_path, 'png_output')
    os.makedirs(png_path, exist_ok = True)
    print(f'Saving images for '+str(row['patient_id']))
    plotting.plot_img(orig_fla_nii, title='fla', output_file=str(Path(png_path, str(row['patient_id'])+'flaa.png')), **plotting_dict)
    plotting.plot_stat_map(seg_fla_nii, orig_fla_nii, output_file=str(Path(png_path, str(row['patient_id'])+'seg.png')), title='Segmentation', **plotting_dict)
