==> code_updated/05_segmap_to_outcome.ipynb <==
import os
import glob
import pandas as pd 
from pathlib import Path
import numpy as np 
from scipy.io import loadmat
from torch.utils.data import Dataset
import json
import nibabel as nib

pd.set_option('display.max_rows', 500)
pd.set_option('display.max_columns', 500)
pd.set_option('display.width', 1000)

# Read in the outcome files: 

os.getcwd()

PATH = Path('/working/lupolab/julia/')

outcome1 = pd.read_csv(Path(PATH, 'cnn_analysis/who_subtype/data/raw_csvs/tcga_subtype_data.csv'))
outcome2 = pd.read_table(Path(PATH, 'cnn_analysis/who_subtype/code/09_createSeparateMasterList_TCIAWHO/gbm_tcga_pub2013_clinical_data.tsv'))
outcome1.shape, outcome2.shape

outcome1 = outcome1.drop(0)
outcome1.head()

outcome1['IDH/codel subtype'].value_counts()

169/(169+263+362)

outcome2.head()

outcome1.columns =['patient_id', '1p19q', 'idh_codel_subtype', 'mgmt_status', 'OS', 'IDH1_drop', 'idh1_mut']

cols_to_drop = [x for x in list(outcome2.columns) if x not in ['Patient ID', 'IDH1 Mutation', 'MGMT Status', 'Overall Survival (Months)']]

outcome2 = outcome2.drop(cols_to_drop, axis = 1)

outcome2.columns = ['patient_id', 'idh_mut', 'mgmg_status', 'OS']

# Create the final outcome

outcome1.columns

outcome1['outcome'], outcome1['cohort'] = ['' for x in outcome1.patient_id], ['' for x in outcome1.patient_id]
outcome2['outcome'], outcome2['cohort'] = ['' for x in outcome2.patient_id], ['TCGA-GBM' for x in outcome2.patient_id]

outcome1 = outcome1.sort_values(by = 'patient_id').reset_index(drop = True)

outcome1.at[:284, 'cohort'] = 'TCGA-GBM'
outcome1.at[284:, 'cohort'] = 'TCGA-LGG'

outcome1['1p19q'].value_counts()

outcome1['idh_codel_subtype'].value_counts()

for idx, row in outcome1.iterrows():
    if row['cohort'] == 'TCGA-GBM' and row['idh_codel_subtype'] == 'IDHwt': 
        outcome1.at[idx, 'outcome'] = 'gbm_idh_wt'
    elif row['cohort'] == 'TCGA-GBM' and row['idh_codel_subtype'] != 'IDHwt': 
        outcome1.at[idx, 'outcome'] = 'as_idh_mut'
    elif row['cohort'] == 'TCGA-LGG' and row['idh_codel_subtype'] == 'IDHwt': 
        outcome1.at[idx, 'outcome'] = 'gbm_idh_wt'
    elif row['cohort'] == 'TCGA-LGG' and row['idh_codel_subtype'] == 'IDHmut-non-codel': 
        outcome1.at[idx, 'outcome'] = 'as_idh_mut'
    elif row['cohort'] == 'TCGA-LGG' and row['idh_codel_subtype'] == 'IDHmut-codel': 
        outcome1.at[idx, 'outcome'] = 'od_1p19codel'

outcome1.outcome.value_counts()

outcome1.outcome.isnull().value_counts()

outcome1.outcome.value_counts()

outcome2.idh_mut.value_counts()

for idx, row in outcome2.iterrows():
    if row['cohort'] == 'TCGA-GBM' and row['idh_mut'] == 'WT': 
        outcome2.at[idx, 'outcome'] = 'gbm_idh_wt'
    elif row['cohort'] == 'TCGA-GBM' and row['idh_mut'] != 'R132H': 
        outcome2.at[idx, 'outcome'] = 'as_idh_mut'
    else: 
        outcome2.at[idx, 'outcome'] = 'gbm_nos'

outcome2.outcome.value_counts()

outcome1.columns

outcome2.columns

outcome2['idh_codel_subtype'] = ['IDHwt'  if x == 'WT' else 'IDHmut-non-codel' if x == 'R132H' else 'gbm_nos' for x in list(outcome2.idh_mut)]

outcomes = outcome1[['patient_id', 'outcome', 'cohort', 'idh_codel_subtype']].append(outcome2[['patient_id', 'outcome', 'cohort', 'idh_codel_subtype']])

outcomes.outcome.value_counts()

outcomes.head()

outcomes.shape

# Merge the outcomes with the segmentation map: 

seg_t1c_df = pd.read_csv(Path(PATH, 'tcia_analysis/datasets/segmentation_t1c_df.csv'))
seg_fla_df = pd.read_csv(Path(PATH, 'tcia_analysis/datasets/segmentation_fla_df.csv'))

seg_t1c_df.shape

seg_fla_df.shape

seg_df = pd.merge(seg_t1c_df, seg_fla_df, how = 'outer', on = ['patient_id', 'cohort', 'patient_exam'], suffixes =['_t1ca', '_fla'])

seg_w_outcome = pd.merge(seg_df, outcomes, on = ['patient_id'], how = 'left')

seg_w_outcome.shape

seg_w_outcome.head()

seg_w_outcome.outcome.value_counts()

seg_w_outcome.loc[seg_w_outcome.outcome.isnull()].patient_id

## Adding in missing patients: 

missing_patients = pd.read_table(Path(PATH, 'cnn_analysis/who_subtype/data/raw_csvs/missing_patients.tsv'))

missing_patients.head()

cols_to_drop = [x for x in list(missing_patients.columns) if x not in ['Patient ID', 'IDH1 Mutation', 'IDH-1P10Q Subtype', 'IDH/codel subtype','IDH status', 'MGMT Status', 'Overall Survival (Months)']]

missing_patients = missing_patients.drop(cols_to_drop, axis = 1)

missing_patients

missing_patients.columns =['patient_id', '1p19q', 'idh_codel_subtype', 'idh_status', 'OS']

missing_patients['outcome'],  missing_patients['cohort'] = ['' for x in missing_patients.patient_id], ['' for x in missing_patients.patient_id]
missing_patients.at[:8, 'cohort'] = 'TCGA-GBM'
missing_patients.at[8:, 'cohort'] = 'TCGA-LGG'

missing_patients

for idx, row in missing_patients.iterrows():
    if row['cohort'] == 'TCGA-GBM' and row['idh_codel_subtype'] == 'IDHwt': 
        missing_patients.at[idx, 'outcome'] = 'gbm_idh_wt'
    elif row['cohort'] == 'TCGA-GBM' and row['idh_codel_subtype'] != 'IDHwt': 
        missing_patients.at[idx, 'outcome'] = 'as_idh_mut'
    elif row['cohort'] == 'TCGA-LGG' and row['idh_codel_subtype'] == 'IDHwt': 
        missing_patients.at[idx, 'outcome'] = 'as_idh_wt'
    elif row['cohort'] == 'TCGA-LGG' and row['idh_codel_subtype'] == 'IDHmut-non-codel': 
        missing_patients.at[idx, 'outcome'] = 'as_idh_mut'
    elif row['cohort'] == 'TCGA-LGG' and row['idh_codel_subtype'] == 'IDHmut-codel': 
        missing_patients.at[idx, 'outcome'] = 'od_1p19codel'

outcomes.head()

outcomes = outcomes.append(missing_patients[['patient_id', 'outcome', 'cohort', 'idh_codel_subtype']])

outcomes['cohort'] =[x+'-nifti' for x in outcomes.cohort]

seg_df.head()

seg_w_outcome = pd.merge(seg_df, outcomes, on = ['patient_id', 'cohort'], how = 'left')

seg_w_outcome.head()

seg_w_outcome.outcome.isnull().value_counts()

seg_w_outcome.outcome.value_counts()

seg_w_outcome.loc[seg_w_outcome.outcome == '']

seg_w_outcome.duplicated().value_counts()

seg_df.cohort.value_counts()

seg_w_outcome.loc[seg_w_outcome.cohort == 'TCGA-LGG-nifti'].idh_codel_subtype.value_counts()

seg_w_outcome.seg_loc_t1ca[0]

import nibabel as nb

nb_file = nb.load(Path(seg_w_outcome.seg_loc_t1ca[0], seg_w_outcome.seg_file_t1ca[0]))

nb_file.get_fdata().shape

nb_file = nb.load(Path(seg_w_outcome.seg_loc_fla[0], seg_w_outcome.seg_file_fla[0]))

nb_file.get_fdata().shape

seg_w_outcome.head()

seg_w_outcome.to_csv(Path(PATH, 'cnn_analysis/who_subtype/data/processed_csvs/seg_map_with_outcome.csv'), index = False)
